feat(runtime): apply basic per section to runtimes
