+++ runtime/common/src/payment.rs
@@ -3,11 +3,7 @@
 
 //! Multiasset related fungibles adapter to allow payments in multiple assets
 
-use frame_support::{
-	sp_runtime::DispatchError,
-	traits::tokens::BalanceConversion,
-	weights::constants::{ExtrinsicBaseWeight, WEIGHT_PER_SECOND},
-};
+use frame_support::{sp_runtime::DispatchError, traits::tokens::BalanceConversion};
 use primitives::{traits::NavProvider, AssetId, Balance};
 use sp_std::marker::PhantomData;
 
@@ -26,22 +22,4 @@ where
 	fn to_asset_balance(balance: Balance, asset_id: AssetId) -> Result<Balance, Self::Error> {
 		NAV::asset_equivalent(balance, asset_id)
 	}
-}
-
-fn base_tx_in_pint() -> Balance {
-	1 / 1000
-}
-
-pub fn basic_per_second() -> u128 {
-	let base_weight = Balance::from(ExtrinsicBaseWeight::get());
-	let base_tx_per_second = (WEIGHT_PER_SECOND as u128) / base_weight;
-	base_tx_per_second * base_tx_in_pint()
-}
-
-pub fn ksm_per_second() -> u128 {
-	basic_per_second() / 50
-}
-
-pub fn dot_per_second() -> u128 {
-	basic_per_second() / 50
 }
